Add concave surface detection, pause fix, frustum far clip plane improvements
Added concave surface detection, allowing it to place points in indents in meshes. Previously, the raycast would ignore the mesh after it touched the mesh once, even if the line exited the mesh and struck the same mesh again due to an indent.
Fixed a race condition causing the pause button to sometimes not work.
Frustum far clip plane was previously a plane (as a true frustum far clip works), but in practice due to how close this clip plane was, it looked strangely "fish eye" like along the edges. Changed to pythagoras to make it look more natural.