Adjusted maximumRaysPerFrame to match existing configs
Due to the concave change plus a previous change which measures each surface hit by the rays instead of just the rays themselves (more surfaces take more time for the raycast), the number of maximumRaysPerFrame required to be playable increased by about a factor of 10. Now internally adjusting by a factor of 10 so that existing configs won't appear broken.